Rename to Core and Extensions
