+++ src-tauri/src/core/helpers.rs
