+++ src-tauri/src/core/helpers/structs.rs
