+++ src-tauri/src/extensions/ugg/runes.rs
@@ -1,4 +1,4 @@
-use crate::{frontend_types::{Active, RuneImages}, shared::helpers};
+use crate::{frontend_types::{Active, RuneImages}, core::helpers};
 
 use super::{structs, constants, json};
